[not verified] move other components
