+++ projects/plugins/jetpack/extensions/blocks/ai-assistant/extensions/multiple-blocks-edition/generate-panel.tsx
@@ -19,10 +19,10 @@ import React from 'react';
  * Internal dependencies
  */
 import { LANGUAGE_MAP } from '../../components/i18n-dropdown-control';
+import { PROMPT_TONES_MAP } from '../../components/tone-dropdown-control';
 import { buildPromptTemplate } from '../../create-prompt';
 import { askQuestion } from '../../get-suggestion-with-stream';
 import Icon from '../../icons/ai-assistant';
-import { PROMPT_TONES_MAP } from '../../tone-dropdown-control';
 
 // Create a Tone Array of objects with `key` and `name` keys
 export const toneOptions = Object.keys( PROMPT_TONES_MAP ).map( key => {
